refactor: 💡 (admin-e2e): Extract getting active session
✅ Closes: ICU-8564
